test(tool-call): fold normalization tests into test_tool.py
Tool, ToolCalls, and to_tool_call all live in dspy/adapters/types/tool.py
so their tests live in a single tests/adapters/test_tool.py file.
